Исправления предупреждений: Конвертации double to float

+++ xrGame/ik/eulersolver.cxx
@@ -141,16 +141,16 @@ void euler_extract(const EulerTableEntry& E, const Matrix R, float vals[3], int 
 			// Family1 cos(v) > 0
 
 			*t[0] = asin1(v);
-			*t[1] = angle_normalize(atan2(y[0], x[0]));
-			*t[2] = angle_normalize(atan2(y[1], x[1]));
+			*t[1] = angle_normalize(atan2f(y[0], x[0]));
+			*t[2] = angle_normalize(atan2f(y[1], x[1]));
 		}
 		else
 		{
 			// Family2 cos(v) < 0
 
 			*t[0] = asin2(v);
-			*t[1] = angle_normalize(atan2(-y[0], -x[0]));
-			*t[2] = angle_normalize(atan2(-y[1], -x[1]));
+			*t[1] = angle_normalize(atan2f(-y[0], -x[0]));
+			*t[2] = angle_normalize(atan2f(-y[1], -x[1]));
 		}
 	}
 	else
@@ -194,8 +194,8 @@ void euler_extract2(const EulerTableEntry& E, const Matrix R, float f1[3], float
 		// Family1 cos(v) > 0
 
 		*t1[0] = asin1(v);
-		*t1[1] = angle_normalize(atan2(y[0], x[0]));
-		*t1[2] = angle_normalize(atan2(y[1], x[1]));
+		*t1[1] = angle_normalize(atan2f(y[0], x[0]));
+		*t1[2] = angle_normalize(atan2f(y[1], x[1]));
 
 		*t2[0] = angle_normalize(M_PI - *t1[0]);
 		*t2[1] = angle_normalize(*t1[1] + M_PI);
