+++ xrGame/ik/jtlimits.h
@@ -40,47 +40,47 @@
 //
 
 // -Pi/2 to  Pi/2 (quadrants IV,I)
-inline double asin1(double x)
+inline float asin1(float x)
 {
 	if (_abs(x) > 1.0f)
 	{
 		//	printf("Domain error in asin1 %lf\n", x);
 		x = (x > 0.f) ? 1.0f : -1.0f;
 	}
-	return angle_normalize(asin(x));
+	return angle_normalize(asinf(x));
 }
 
 //  Pi/2 to -Pi/2 (quadrants II,III)
-inline double asin2(double x)
+inline float asin2(float x)
 {
 	if (_abs(x) > 1.0f)
 	{
 		//	printf("Domain error in asin2 %lf\n", x);
 		x = (x > 0) ? 1.0f : -1.0f;
 	}
-	return angle_normalize(M_PI - asin(x));
+	return angle_normalize(M_PI - asinf(x));
 }
 
 //  0 to Pi   (quadrants I,II)
-inline double acos1(double x)
+inline float acos1(float x)
 {
 	if (_abs(x) > 1.0f)
 	{
 		//	printf("Domain error in acos1 %lf\n", x);
 		x = (x > 0) ? 1.0f : -1.0f;
 	}
-	return angle_normalize(acos(x));
+	return angle_normalize(acosf(x));
 }
 
 //  Pi to 2Pi (quadrants III,IV)
-inline double acos2(double x)
+inline float acos2(float x)
 {
 	if (_abs(x) > 1.0f)
 	{
 		//	printf("Domain error in acos2 %lf\n", x);
 		x = (x > 0) ? 1.0f : -1.0f;
 	}
-	return angle_normalize(-acos(x));
+	return angle_normalize(-acosf(x));
 }
 
 enum
