+++ xrGame/ui/UIMainIngameWnd.cpp
@@ -278,8 +278,8 @@ void CUIMainIngameWnd::Draw()
 	if (!IsGameTypeSingle())
 	{
 		float luminocity = smart_cast<CGameObject*>(Level().CurrentEntity())->ROS()->get_luminocity();
-		float power = log(luminocity > .001f ? luminocity : .001f) * (1.f /*luminocity_factor*/);
-		luminocity = exp(power);
+		float power = logf(luminocity > .001f ? luminocity : .001f) * (1.f /*luminocity_factor*/);
+		luminocity = expf(power);
 
 		static float cur_lum = luminocity;
 		cur_lum = luminocity * 0.01f + cur_lum * 0.99f;
